+++ Module.php
@@ -34,13 +34,26 @@ class Module extends \humhub\components\Module
         return $verifySpace;
     }
 
-    public function getMaxNumber()
+    public function getMaxUsers()
     {
-        $maxNumber = $this->settings->get('maxNumber');
-
+        $maxNumber = $this->settings->get('maxUsers');
+        
         if (empty($maxNumber)) {
-            return [];
+            $maxNumber = '25';
         }
+        
+        return $maxNumber;
+    }
+    
+    public function getMaxSpaces()
+    {
+        $maxNumber = $this->settings->get('maxSpaces');
+        
+        if (empty($maxNumber)) {
+            $maxNumber = '25';
+        }
+        
+        return $maxNumber;
     }
 
     public function getUserIcon()