+++ docs/CHANGELOG.md
@@ -2,6 +2,8 @@ Changelog
 =========
 ### (TBA)
 - Fix: Profile Image upload/edit
+- Enh: Fix maximum number of allowed accounts
+- Enh: add setting: maximum number of allowed spaces
 
 ### 1.0.0-beta.3 (4/20/2023)
 - Enh: People and spaces page